# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equilibrium blockchain node RPC + block explorer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChainStatus {
  height: number;
  latestHash: string;
  latestTimestamp: number;
  totalTxCount: number;
  mempoolSize: number;
  mempoolPressure: number;
  validatorCount: number;
  cumulativeWork: number;
  lastResidual: number;
  avgBlockTime: number;
  tps: number;
}

export interface BlockStat {
  height: number;
  txCount: number;
  residual: number;
  mempoolPressure: number;
  timestamp: number;
}

export interface Peer {
  peerId: string;
  address: string;
  latencyMs: number;
  height: number;
  connected: boolean;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Transaction {
  hash: string;
  from: string;
  to: string;
  amount: number;
  fee: number;
  nonce: number;
  /** @nullable */
  blockHash: string | null;
  /** @nullable */
  blockHeight: number | null;
  timestamp: number;
  status: TransactionStatus;
}

export interface Block {
  hash: string;
  height: number;
  prevHash: string;
  merkleRoot: string;
  timestamp: number;
  nonce: number;
  difficulty: number;
  residual: number;
  recursionDepth: number;
  coinbaseReward: number;
  miner: string;
  txCount: number;
  transactions: Transaction[];
}

export interface BlockPage {
  blocks: Block[];
  total: number;
  page: number;
  limit: number;
}

export interface SignedTxInput {
  from: string;
  to: string;
  amount: number;
  fee: number;
  nonce: number;
  signature: string;
  publicKey: string;
}

export interface BroadcastResult {
  txHash: string;
  status: string;
}

export interface AddressInfo {
  address: string;
  balance: number;
  nonce: number;
  txCount: number;
  transactions: Transaction[];
}

export interface Mempool {
  count: number;
  pressure: number;
  transactions: Transaction[];
}

export interface ApiError {
  error: string;
}

export type ListBlocksParams = {
page?: number;
limit?: number;
};