# lib/api-zod/src/generated/types/addressInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equilibrium blockchain node RPC + block explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface AddressInfo {
  address: string;
  balance: number;
  nonce: number;
  txCount: number;
  transactions: Transaction[];
}