# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equilibrium blockchain node RPC + block explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  error: string;
}