# lib/api-zod/src/generated/types/blockPage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equilibrium blockchain node RPC + block explorer API
 * OpenAPI spec version: 0.1.0
 */
import type { Block } from './block';

export interface BlockPage {
  blocks: Block[];
  total: number;
  page: number;
  limit: number;
}