# lib/api-zod/src/generated/types/signedTxInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equilibrium blockchain node RPC + block explorer API
 * OpenAPI spec version: 0.1.0
 */

export interface SignedTxInput {
  from: string;
  to: string;
  amount: number;
  fee: number;
  nonce: number;
  signature: string;
  publicKey: string;
}